+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Optional `freeze` feature and `--freeze` flag: freezes the screen while a selection is in progress.
+
 ## [0.6.1] - 2026-03-24
 
 ### Changed
+
 - Fixed panic on hyperland by @id3v1669
 - Set the namespace to osk
 - make dimension_or_output also show text